requirements.txt
torch
torchvision
numpy
scipy
pytorchcv
pillow
